+++ template/runtime/src/configs/xcm_config.rs
@@ -394,7 +394,7 @@ impl pallet_xcm::Config for Runtime {
   type TrustedLockers = ();
   type UniversalLocation = UniversalLocation;
   type Weigher = FixedWeightBounds<UnitWeightCost, RuntimeCall, MaxInstructions>;
-  type WeightInfo = pallet_xcm::TestWeightInfo;
+  type WeightInfo = crate::weights::pallet_xcm::SubstrateWeight<Runtime>;
   // ^ Disable dispatchable execute on the XCM pallet.
   // Needs to be `Everything` for local testing.
   type XcmExecuteFilter = Nothing;