+++ template/runtime/src/weights/mod.rs
@@ -27,6 +27,7 @@ pub mod pallet_asset_registry;
 pub mod pallet_axial_router;
 pub mod pallet_governance;
 pub mod pallet_staking;
+pub mod pallet_xcm;
 
 pub mod pallet_tmc;